Moved BaseResult and StoredResult to a new results folder
