+++ maboss/__init__.py
@@ -1,6 +1,7 @@
 from .network import *
 from .simulation import *
 from .result import *
+from .results import *
 from .gsparser import load
 from .server import MaBoSSClient
 from .upp import UpdatePopulation
