+++ maboss/results/__init__.py
@@ -1,0 +1,1 @@
+from .storedresult import StoredResult
